Fix transition matrix and visualisation functions
The transition_matrix function now accepts the build parameter, which
is used when it calls the get_states function.

The edge between the "S" node and the first node in the pattern was
labelling the edge with first letter in the alphabet. It now just
returns the number of sequences entering the first node.